+++ geotrek/api/v2/serializers.py
@@ -512,6 +512,7 @@ if 'geotrek.tourism' in settings.INSTALLED_APPS:
 
     class TouristicEventSerializer(TouristicModelSerializer):
         organizers = serializers.SerializerMethodField()
+        organizer = serializers.SerializerMethodField()
         organizers_id = serializers.SerializerMethodField()
         attachments = AttachmentSerializer(many=True, source='sorted_attachments')
         url = HyperlinkedIdentityField(view_name='apiv2:touristicevent-detail')
@@ -553,6 +554,9 @@ if 'geotrek.tourism' in settings.INSTALLED_APPS:
                 map(lambda org: org.label, obj.organizers.all())
             )
 
+        # for retrocompatibility of API
+        get_organizer = get_organizers
+
         def get_organizers_id(self, obj):
             return ", ".join(
                 map(lambda org: org.label, obj.organizers.all())
@@ -565,7 +569,7 @@ if 'geotrek.tourism' in settings.INSTALLED_APPS:
                 'booking', 'cancellation_reason', 'cancelled', 'capacity', 'cities',
                 'contact', 'description', 'description_teaser', 'districts', 'duration',
                 'email', 'end_date', 'end_time', 'external_id', 'geometry', 'meeting_point',
-                'meeting_time', 'name', 'organizers', 'organizers_id', 'participant_number', 'pdf', 'place',
+                'meeting_time', 'name', 'organizers', 'organizer', 'organizers_id', 'participant_number', 'pdf', 'place',
                 'portal', 'practical_info', 'provider', 'published', 'source', 'speaker',
                 'start_time', 'structure', 'target_audience', 'themes', 'type',
                 'url', 'uuid', 'website'